Fix logic in `Transaction.rollback()`
It's common to have code that looks like the following:

```dart
try {
  ...
  transaction.commit();
} catch (error) {
  transaction.rollback();
}
```

This code is prone to errors - if the commit fails (and
throws an exception), then the attempt to call `rollback()`
will throw a secondary error (transaciton has already been
committed),  thus masking the real error.

This PR fixes that logic.