Add `DatastoreDB.lookupValue()`
`DatastoreDB.lookup()` returns null rows in its
results, which is confusing - the caller calls `lookup(keys)`,
none of which are successfully looked up, and instead of getting
an empty list back, they get a list of nulls.

Since changing that behavior would be an API breaking
change, and since a very common use-case is calling
`lookup(<Key>[key])`, this change adds a new API method for
looking up a single key. This new method will throw if the
lookup was unsuccessful.