Fix `Page.next()` when `Page.isLast` is `true`.
Prior to null-safety migration `Page.next()` would return
`Future.value(null)` when `Page.isLast` is `true`. This was not
correctly migrated to null-safety as a result code-bases that are fully
migrated to null-safety would get a runtime error. Code-bases that were
not migrated to null-safety or only partially migrated to null-safety
would get `null` in a result that is not nullable.

This affects Pub-Sub and GCS clients users, if and only if they call
`Page.next()` when `Page.isLast` is `true`, and their code-bases are
not migrated to null-safety. The Cloud Datastore client already throws
an `Error`, so no-one is broken by changing which `Error` is thrown as
we generally do not make promises about which `Error` is thrown.